pgy3 form changed to RotationForm
bit less pgy3 focused. i think there could be more done to make it better but, if it's needed we'll do it.
